Download uamp

+++ backend/MpumalangaAssetManagement/MAM.DataAccess/Tables/StrategicAssessment.cs
@@ -9,9 +9,9 @@ namespace MAM.DataAccess.Tables
         public int Id { get; set; }
         public int UserImmovableAssetManagementPlanId { get; set; }
         public string PostDescriptionTitle { get; set; }
-        public decimal? AllocatedSpace { get; set; }
+        public double? AllocatedSpace { get; set; }
         public string SurplusShortageAccommodation { get; set; }
-        public int? PercentageUtilised { get; set; }
+        public double? PercentageUtilised { get; set; }
         public decimal? FbpLevel { get; set; }
         public decimal? FbpQuantity { get; set; }
         public decimal? FbpNorm { get; set; }
